requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# Google AI / Gemini
google-generativeai>=0.3.0

# OpenAI / GPT
openai>=1.0.0

# ElevenLabs Text-to-Speech
elevenlabs>=0.2.0

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3

# Environment Variables
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Async Support
aiofiles>=23.0.0

# HTTP Requests
requests>=2.31.0

# Logging
structlog>=23.0.0

# Type Hints
typing-extensions>=4.8.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
